fix pathfinding for some circumstances
change Bos.OS.Cmd.exists to Bos.OS.File.exists; Bos.OS.Cmd.exists will
just search the PATH and give an answer directly, rather than giving us
the correct answer for *which* directory contains the file.

TODO this probably gives the wrong answer when there is a
non-executable file with the same name as the fuzzer in a directory in
the PATH which occurs before the directory containing the binary.