try harder to avoid CPU detection collisions
Only launch the "first" afl-fuzz instance outside of `fill_cores` when
the user has explicitly requested only a single core.

This change in program flow is meant to avoid collisions between the
first afl-fuzz instance's internal invocation of the `afl-gotcpu` logic,
and our invocation for purposes of discovering how many `afl-fuzz`
instances to launch.  If more than one process is trying to use the
pre-emption metrics at the same time, both will come to the conclusion
that many CPUs are more subscribed than they are.