Make bun compatible with both Dune 1.11 and 2.x
Otherwise trying to install bun would downgrade dune to 1.11.4,
which would then fail to build my project which uses Dune 2.x.

See https://github.com/ocaml/dune/issues/2912, bun was relying on
a dune feature that was part of a breaking change in 2.0.

Move the new target (creation of file) out of the alias into a separate
rule, which then works on both old and new dune.

Dune told me that the features used in the dune file now require 1.11,
so update dependencies accordingly.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>